requirements.txt
python-dotenv
PyGithub
GitPython
pandas
cliffs_delta
scipy